Scaffold MCP endpoint
