+++ apps/docs/content/docs/docs/mcp.mdx
@@ -1,0 +1,9 @@
+---
+title: Model Context Protocol
+description: Kibo UI supports the Model Context Protocol (MCP) for model-driven development.
+icon: Code
+---
+
+Kibo UI supports the Model Context Protocol (MCP) for model-driven development.
+
+## Installation
