+++ apps/docs/content/docs/docs/meta.json
@@ -9,6 +9,7 @@
     "---Usage---",
     "setup",
     "usage",
+    "mcp",
     "troubleshooting",
     "---Contributing---",
     "how-to-contribute",
