+++ apps/docs/package.json
@@ -55,6 +55,7 @@
     "@t3-oss/env-nextjs": "^0.13.6",
     "@tailwindcss/postcss": "^4.1.8",
     "@vercel/analytics": "^1.5.0",
+    "@vercel/mcp-adapter": "^0.2.4",
     "date-fns": "^4.1.0",
     "fumadocs-core": "15.3.3",
     "fumadocs-mdx": "11.6.4",
